+++ lib/src/websocket_base/services/websocket_handler_io.dart
@@ -132,6 +132,7 @@ class WebsocketHandlerIo<T, Y> implements IWebSocketHandler<T, Y> {
       _setInitPing();
       if (isConnected) {
         await _connectionSuccessful();
+        _isConnectionAlivePing();
         return true;
       }
       await _connectionUnsuccessful();
@@ -164,8 +165,6 @@ class WebsocketHandlerIo<T, Y> implements IWebSocketHandler<T, Y> {
     }
     _webSocket = await io.WebSocket.connect(connectUrl)
         .timeout(Duration(milliseconds: _timeoutConnectionMs));
-    _webSocket?.pingInterval = const Duration(milliseconds: 500);
-    _isConnectionAlivePing();
     return true;
   }
 
@@ -200,7 +199,7 @@ class WebsocketHandlerIo<T, Y> implements IWebSocketHandler<T, Y> {
     await _toServerMessagesSub?.cancel();
     _toServerMessagesSub = outgoingMessagesStream
         .takeWhile((_) => socketState.status == SocketStatus.connected)
-        .listen(_addMessageToSocketOutgoingInternal);
+        .listen((e) => _addMessageToSocketOutgoingInternal(e, true));
   }
 
   ///
@@ -227,33 +226,45 @@ class WebsocketHandlerIo<T, Y> implements IWebSocketHandler<T, Y> {
     if (isPing) {
       _startPingRequest();
     }
-    if (data is Y) {
-      final outJsonMsg = _messageProcessor.serializeMessage(data);
 
-      if (!isPing || !_skipPingMessages) {
-        // This controller's stream is listened by [_listenMessagesToServer()]
-        _outgoingMessagesController.add(outJsonMsg);
-        _debugEventNotificationInternal(
-          SocketLogEventType.toServerMessage,
-          'to server',
-          data: outJsonMsg.toString(),
-        );
-      } else {
-        // Generally this branch is for ping messages:
-        _addMessageToSocketOutgoingInternal(outJsonMsg);
-      }
+    Object? dataToSend;
+    if (data is Y) {
+      dataToSend = _messageProcessor.serializeMessage(data);
     } else {
-      // Generally this branch is for ping messages:
-      _addMessageToSocketOutgoingInternal(data);
+      dataToSend = data;
+    }
+    if (dataToSend == null) {
+      _debugEventNotificationInternal(
+        SocketLogEventType.warning,
+        'Trying to send NULL data!',
+      );
+      return;
+    }
+    if (!isPing) {
+      // This controller's stream is listened by [_listenMessagesToServer()]
+      _outgoingMessagesController.add(dataToSend);
+    } else {
+      if (_skipPingMessages) {
+        _addMessageToSocketOutgoingInternal(dataToSend, false);
+      } else {
+        _outgoingMessagesController.add(dataToSend);
+      }
     }
   }
 
   /// Sending to server platform implementation:
-  void _addMessageToSocketOutgoingInternal(Object? input) {
+  void _addMessageToSocketOutgoingInternal(Object? input, bool notify) {
     try {
       if (input == null) {
         throw ArgumentError.notNull('[_addMessageToSocketOutgoingInternal] '
             'sent data must not be NULL!');
+      }
+      if (notify) {
+        _debugEventNotificationInternal(
+          SocketLogEventType.toServerMessage,
+          'to server',
+          data: input.toString(),
+        );
       }
 
       _addToSocketPlatform(input);
@@ -394,11 +405,7 @@ class WebsocketHandlerIo<T, Y> implements IWebSocketHandler<T, Y> {
 
   void _isConnectionAlivePing({String? message}) {
     if (_checkPlatformIsConnected('Ping socket.')) {
-      if (!_skipPingMessages && _messageProcessor.pingServerMessage is Y) {
-        sendMessage(_messageProcessor.pingServerMessage as Y);
-      } else {
-        _sendMessageInternal(_messageProcessor.pingServerMessage, true);
-      }
+      _sendMessageInternal(_messageProcessor.pingServerMessage, true);
 
       final msg = 'Ping socket. Status: ${socketState.status.value}.';
       _debugEventNotificationInternal(