+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # 0.0.2-dev
 
 - refactor: rename library to `websocket_universal`
+- exported required classes
+- added example
 
 # 0.0.1-dev
 
 - added websocket implimentations for io and html
+