+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.3-dev
+
+- now supports Byte array messages
+
 # 0.0.2-dev
 
 - refactor: rename library to `websocket_universal`